refactor(sdk): Change waveform to be a list of values between 0 and 1
Most clients will probably work with values between 0 and 1 and need to
convert it just to send it, so we can move that conversion into the SDK.

This is also more forwards-compatible, because MSC3246 now has a
different max value for the amplitude, so when this becomes stable, the
only change needed will be in the SDK.

+++ crates/matrix-sdk/src/attachment.rs
@@ -67,7 +67,9 @@ pub struct BaseAudioInfo {
     /// The file size of the audio clip in bytes.
     pub size: Option<UInt>,
     /// The waveform of the audio clip.
-    pub waveform: Option<Vec<u16>>,
+    ///
+    /// Must only include values between 0 and 1.
+    pub waveform: Option<Vec<f32>>,
 }
 
 /// Base metadata about a file.
